change alert position

+++ src/app/auth/signup/signup.component.ts
@@ -24,6 +24,7 @@ export class SignupComponent implements OnInit {
   // && this.myForm.value.uemail == "" && this.myForm.value.upass=="" && this.myForm.value.upassconfirm==""
   
   onSubmit(){
+    console.log(this.myForm)
     if(this.myForm.value.uname =="" && this.myForm.value.uemail == "" && this.myForm.value.upass=="" && this.myForm.value.upassconfirm=="" ){
       // alert("please fill all values");
       this.show2=true;
@@ -33,7 +34,10 @@ export class SignupComponent implements OnInit {
     }
     else{
       if(this.myForm.value.upass==this.myForm.value.upassconfirm &&
-        this.myForm.value.uname.length > 0 && this.myForm.value.uemail.length > 0 && this.myForm.value.upass.length > 0 && this.myForm.value.upassconfirm.length > 0){
+        this.myForm.value.uname.length > 0 && this.myForm.value.uemail.length > 0 &&
+         this.myForm.value.upass.length > 0 && this.myForm.value.upassconfirm.length > 0
+         && this.myForm.controls.uname.valid && this.myForm.controls.uemail.valid && this.myForm.controls.upass.valid && this.myForm.controls.upassconfirm.valid
+         ){
         console.log(this.myForm.value)
         console.log(this.myForm.touched)
         localStorage.setItem("data",JSON.stringify(this.myForm.value));
